requirements.txt
python-dotenv
openai
transformers